Final approach - identify each item subscale for each age, but choose a number of items and threshold that keeps false positives low.
In this version, I order the items by how well they predict poor performance on that item type for that age group. Then for each age group and item type, I consider all lengths and threshold scores to calculate a) the detection rate for at risk (20th percentile) on that item type and age, and b) if the detection rate is perfect, I calculate the false positive rate for that list length and threshold.

Then, using my eyeballs, I choose a length and threshold that keeps the false positive rate below 10% or as low as possible (sometimes, if one extra item improves the score, I might do that).